Move BaremetalInventory, HostsTable and ClusterWizardToolbar to clusterWizard
